pyproject.toml
[tool.poetry]
name = "imposter-game"
version = "0.1.0"
description = ""
authors = ["Eugene <giveuper39@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
aiogram = "^3.13.1"
peewee = "^3.17.7"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
black = "^24.10.0"

[tool.black]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D106"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
